pbio/parent: Link objects so they can stop each other.
Until this commit, a servo could not start if a drive base was using it,
and a dc motor could not start if a servo was using it. This was managed
by several (different per type) claim flags, returning EBUSY if it was
in use.

This commit introduces a generic pbio_parent type, which makes a child
object aware of its parent. Each type can provide the required stop
function (all with the same signature), so that any lower level device
(like the dcmotor) is able to escalate a stop command all the way to the
highest level (servo, then drivebase).

A parent is assigned to the child when the parent object is initialized.

In the future, we can also use this when servos can be in use by other
high-level classes, like a class that syncs N motors. It can also be
extended indefinitely, with higher abstractions above a Drive Base.

This also lets us do several other cleanups. For example, now the
Motor.stop() method is truly the same (inherited) as DCMotor.stop(),
since it is now able to stop the servo control loop.