pbio/servo: Register servos for updates.
Instead of setting a connected flag, explicitly register servos for getting control loop updates.

Also provide a single entry point to update all servos instead of having this logic partially managed by a global process.

Also make the update function for a single servo static and move it closer to the other update and register logic. The update itself is not changed.