pbio/parent: Include clear option in stop call.
Use a bool instead of a separate function to clear the parent. This lets us propagate it to the highest level, just like the stop call.

This avoids partially broken chains of hierarchy, and it simplifies the global reset: we just have to stop/clear the dcmotors, and everything else just gets propagated.