pbio/dcmotor: Simplify user call api.
The option to stop a parent device (introduced previously to
save a few near-duplicate functions) make the dcmotor api
a bit more confusing than it needs to be.

This partially reverts that change and solves it differently, by
making it unambigous which pbio call should be used for user
calls from the MicroPython module.