pbio/servo: Don't allow new commands if not set up.
In practice, this means we don't allow new commands if a control update has failed. This happens when a motor was unplugged.

Since we have no way of knowing what the position is once it is plugged in again, we require that the setup is run again. In MicroPython that means the user will have to initialize the motor object again if they want to use it after unplugging it.