pybricks.common.Motor: Revert ignoring of errors.
This partially reverts 150423664999e98618ad2fb0c61e7e6e1fd2a8cf.

Since motor errors are now being handled properly, we can check if something is wrong when control stops.

If so, raise an error. In practice, this only happens when a cable is pulled while the motor is running, so just raise that one.