pbio/drivebase: Fix order of passive actuation.
If passive actuation is required at the end of a
maneuver (brake or coast), this stops control
automatically.

Changing integration states at that point is
not a valid operation, so we need to put this
after the return on passive actuation.

This fixes coast and brake not working on
drivebases.