pbio/drivebase: Fix near-zero turns.
In a tank drive, when both motors are nearly at the same speed, the
difference controller turns by only 1 or 2 degrees, at an equivalently
low speed. This can take longer than the actual maneuver of interest,
which is not what should happen.

Since the sum_rate and dif_rate are used as upper bounds in the
drivebase control (the angles always lead), we can use the highest
bound for both. The speed for the shorter maneuver gets automatically
reduced anyway by stretching its trajectory. This avoids numerical issues
with speeds in the order of 1 degree per second.