pybricks.pupdevices.Motor: Fix wait for sync.
Previously, it was the setup function that could
return  PBIO_ERROR_AGAIN on first use, but this
has recently moved to the servo getter. We need
to update the MicroPython class initializer as well.

This fixes programs crashing (raising OSError) if
they try to initialize a motor right after boot or
right after plugging it in.