pbio/drivebase: Don't use double torque limit for heading.
There is no good reason for this, so remove it. The logic for it
given in the comments does not apply (and hurts performance)
in case of in-place turns.

If we want the other wheel to stop, we need to ensure that
both controllers pause properly if one of them stalls.