pbio/control: Use correct time for hold.
This makes the hold trajectory account for previously
paused time.

This is not technically *necessary* because the reference is
just a horizontal line (independent of time), but it makes
the datalog time plot much easier to follow, as maneuver
times are now always positive and starting from 0 as they
should be.

Before this fix, they could be negative. Despite having no
impact on control, this was masking another recently-fixed
synchronization bug, so it's probably worth the extra couple
of cycles.