platform/motors/settings: Set ki based on saturation time.
Integral control is used to offset constant disturbances. Instead
of setting a somewhat arbitrary value for each motor, set it such
that integral control reaches its maximum after a few seconds.

This way, the behavior is about the same for all motors - after a
second or two it is able to overcome the highest loads if they
occur.

Fixes https://github.com/pybricks/support/issues/551. Here, a motor
needed nearly a minute to reach peak torque, which is too slow.