pbdrv/bluetooth: make advertising state consistent
the advertising state `advertising_now` needs to be changed when advertising for broadcasts and connecting to code.pybricks.com.
It also turned out that by changing the maximum number of peripheral devices to two for the primehub removed the implicit advertisement stop after connecting to code.pybricks.com. This is also fixed by an expicit advertisement stop.