github: fix build each commit for pull requests
The `pull_request_target` trigger causes github actions to check out
the master branch instead of the pull request branch. This is done
for security reasons since `pull_request_target` is allowed access to
secrets.

This modifies the build-each-commit.py script to be able to run without
secrets so that it can be used for pull requests and adds a new
conditional in the workflow to handle this case.

This should also fix building pull requests that are not based on the
master branch.

Also flush the print() so that the message is printed before the output
of the make command.