drv/usb/stm32: implement battery charger detection
This adds a feature to detect the battery charging capabilities when
something is connected to the USB port. Since all hubs so far are
self-powered, peripheral-only USB ports, the VBUS line is muxed as a
GPIO. Then we use the interrupt on this GPIO to detect when something
with power is plugged in or removed.

When something is plugged in, it triggers the battery charger detection
algorithm to see what sort of charging capabilities the port has. This
has been exposed in the MicroPython API for informational purposes.

Future changes will use this to actually enable/disable charging and
set the max charging current.