pbdrv/charger: Fix fault detection.
The existing computation is wrong in several ways:

For example, assume that CHG is nominally rising, like this:

00000000011111111

a) A fault would be registered on every transition.

In the example, the buffer reads 00011 or 00011 at some point, both of which incorrectly trigger a fault.

b) In nominal operation, the state is wrong just prior to the fault trigger:

In the example, when the buffer is 00001, the computation still says low, when it has already risen.

In this commit we look at the number of transitions instead.