drv/{charger,usb}: separate USB from charger
This does a better job isolating the drivers as abstractions of only
the specific chip/subsystem that they encapsulate and lets the pbsys
layer do the gluing together instead.

This also fixes a crash introduced in commit 472eb7dc "drv/charger_mp2639a:
implement ISET control" that was caused by pbdrv_charger_set_usb_type()
accessing an initialized pointer because the USB driver was calling
it before the charger driver was done initializing.