bricks/virtualhub: add python backend
This embeds Python into the virtualhub to provide hardware simulation
that is implemented in Python. This works by providing a module name
via the VIRTUALHUB_MODULE environment variable. The module must contain
a VirtualHub class that implements the required methods.

A default VirtualHub implementation that does nothing is provided to
document the required methods and a VirtualHub using turtle graphics
is supplied to demonstrate how to provide an implementation that
actually does something.

    PYTHONPATH=tests VIRTUALHUB_MODULE=virtualhub.turtle ./bricks/virtualhub/build/virtualhub-micropython