pybricks/util_pb/pb_imu: Disable high pass filter.
Although it is nice that the moving average is subtracted automatically,
it makes the default behavior harder to understand. Since it is easy
enough to keep track of the average and subtract it in the user script,
the builtin filter is now disabled.

This allows for more generic use cases, and makes it consistent with
other firmwares for these hubs.