drv/motor: remove pbdrv_motor_get_id()
The motor driver is supposed to be and abstraction of the motor driver
chip (h-bridge). The motor type id is unrelated to this - it is a
function of the I/O port the motor is attached to since the port is
responsible for detecting the type ID of the motor.

The functionality is moved to pb_device_setup_motor() since that is
where the port is configured for the expected motor type on platforms
that don't have fully automatic configuration.