src/math: fix pbio_math_clamp()
The pbio_math_clamp() function introduce in a54f7dee "pbio/math: Add
clamp function." contained an error. This adds a unit test to discover
the error and fixes the error. Also add doc comment explaining the rules
of the function and use assert to express assumptions about the range
of the arguments.