drv/motor_driver: change PBDRV_MOTOR_DRIVER_MAX_DUTY
This changes PBDRV_MOTOR_DRIVER_MAX_DUTY from 10000 to 1000. This saves
us from having to divide by 10 in motor_driver_hbridge_pwm. The value
was originally 10000 because that was the PWM period, but since we
changed the PWM frequency by a factor of 10 from ~1kHz to ~10kHz, the
period was reduced by an equal factor.

Other drivers already use PBDRV_MOTOR_DRIVER_MAX_DUTY to scale the duty
and don't require modification.