pybricks.robotics.DriveBase: Refactor settings() method.
The settings() method now calls the corresponding pbio method
directly as it should, with no logic handled in the MicroPython method.

This is achieved by letting the default speed be a generic control setting
instead of a drivebase-specific setting.

Also remove the not-busy assertion that we missed to delete when we
did that for all other control settings.