pbio/doc/control: Simplify time based trajectories.
For angle based maneuvers, we ended up simplying
the result by allowing w3 to be either 0 or w_target,
rather than allowing an arbitrary end speed.

This commit applies the same simplification to the
time based trajectories to keep things consistent.
It also reduces complexity in the implementation
without sacrificing any functionality, because the API
will only allow the aforementioned values for w3.