pbio/trajectory: Simplify trajectory extension logic.
This brings the logic for "extended" trajectories into the control
code instead of in the trajectory code. Instead of explicitly having
a somewhat opaque call to extend it, we calculate the shifted
starting point and re-calculate the trajectory from there.

Specific issues of angle vs. time trajectories are dealt with in the
appropriate control handler instead of having various if/else calls
inside the trajectory generators. This results in slightly more code
but it is much easier to follow.