pbio/dcmotor: fix pbio_dcmotor_stop_all()
In 10e4dbc we moved checking the I/O device capabilities for dcmotor
to pb_device_setup_motor(). This caused pbio_dcmotor_stop_all() to
turn off outputs for all devices instead of just motors which causes
sensors that require DC power to stop functioning.

This fixes the breakage by checking the device ID to decide if we should
turn off the motor driver or not.