pbio/uartdev: Fix Move Hub motor suddenly starting.
When a powered powered up sensor like a color sensor is
plugged in, it turns on the power for that port. On Move
Hub, ports A and B are not ioports, which offsets the driver
index by two. This caused motors A and B to start suddenly
if a color sensor was plugged into port C or D, respectively.

Pending a bigger overhaul of the device drivers, we can patch
the port index for now.