pbio/servo: Drop run_until_stalled.
Since the end-user API call is blocking, we don't need a special
case for it in pbio. We can drop this to further simplify and refactor
pbio in subsequent commits, since run_until_stalled was often
the odd one out.

At the module level, the implementation hardly changes. Instead of
checking for a done-flag that is set by pbio when stalled, we check for
the stalled flag directly. So this does not pose any additional restrictions
when switching to fully async later on.