pybricks.common.Motor: Drop busy() and stalled().
This reverts commit 73a265cbb70a73d5f6826aedb475c2a5071dd8ac.

Since ccc31b965f431a932d54f698a4e67ec60dd33e76, the "busy" terminology
makes a bit less sense. It is implemented as "not on target", but this
is not always the same as "busy".

Since we have not released this (not even in a beta release), we can
drop it so we have a bit more time to settle on the right change.