pbio/control: Fix start time of hold control.
In pbio_control_start_hold_control, the function
pbio_control_get_ref_time could be called when
no controller is active. For example, when holding
from the coast state.

This would lead to an undefined starting time. Since
the hold trajectory is just a stationary line it would still
give the correct speed and position, but the maneuver
time plot in the data logs looks confusing.

This fixes it by always returning the real time when no
controller is active.