drv/virtual: implement CPython exception handling
This adds a mechanism for handling exceptions in CPython. Since each
callback into CPython could raise an exception, we need to check for
this on each time we don't use PyRun_SimpleString (and we should stop
using this at all eventually).

The MicroPython implementation is updated to pass SystemExit and
KeyboardInterrupt exceptions from CPython to MicroPython. Other
unhandled CPython exceptions will schedule a CPythonError in
MicroPython. This should prevent MicroPython from running on when
there was an unhandled exception in CPython.

Making use of the SystemExit forwarding, we can also fix an issue in
the virtualhub.turtle implementation where closing the window caused
things to go haywire.