pbio/servo: add proper config options
This adds a new PBIO_CONFIG_SERVO option to enable the PBIO servo
module. It also gets some child options to selectively enable supported
servo types. Previously, this was using driver config to try to guess
what should be enabled, but this makes it more difficult to add virtual/
test platforms since each new driver config option would have to be
added in many places throughout the code.

While we are touching this, move the servo setting file out of the
platform directory (the platform directory is for platforms where the
Pybricks firmware runs, so unless we are making firmware that runs on
motors, it doesn't make sense to have motor code there). A new motor
directory is started in the pbio/src/ folder. Eventually, other motor-
related source code files should be moved there as well for better
organization.