drv/ioport/virtual: new virtual I/O port driver
This adds a new virtual I/O port driver for virtual hubs. This one is
a bit different from the other drivers since the current I/O port and
I/O device implementation is a bit of a hack. To make it work we have
to recreate the pbio_iodev_t structure in CPython and pass it to the
PBIO runtime.

The led and motor_driver CPython implementations are also reworked so
that we don't have to subclass them. This lets us use the new drivers
in the turtle graphics platform to simulate a motor.