drv/clock/virtual: new virtual clock driver
This implements a new clock driver to work with virtual hubs. It is
similar to the linux clock driver in that a signal is used to interrupt
the main thread.

A WallClock implementation is also provided to make it easy to create
a virtual hub that works in "real" time.