bricks/virtualhub/mp_port: make polling more efficient
This fixes up the MICROPY_VM_HOOK_LOOP and MICROPY_EVENT_POLL_HOOK to
be more efficient. In the former, we should not be checking the
MicorPython scheduler. In the latter, we add a way to efficiently wait
for a signal to wake up the thread but allow sleeping if there really
isn't anything to do.

This greatly reduces the CPU usage for "realtime" virtual hubs.