pbio_virtual.drv.clock: new CountingClock class
This adds a new CountingClock class that allows running the virtual
clock as fast as possible. This is useful for testing/simulation where
we don't want to wait in real time for tests that depend on the passage
of time.

The class also contains several features to help prevent accidentally
developing code that depends on a "perfect" clock.