drv/virtual/simulation: Rework simulation timing.
This updates the simulation so that both the motor driver or the counter can push it onwards. This way it can work even in open loop, by triggering a simulation update when the counter is read.

This is several times faster than polling since it skips unecessary calls, particularly for motors that are not doing anything. It is also ensures the model synchronizes with the data.

Also - unlike the motor driver calls, counter calls do not get time stamps with it, so we have to pass the clock to get these. For consistency, we should either drop it from the motor drivers, or add it to the counter driver.