tests/motors/run_test: Allow plotting of old data.
Sometimes, we just want to plot previously generated data
without running the experiment again. We already archived
the rendered results, but this lets you regenerate the interactive
plot so you can still zoom in or change details in the graph.

To use it, just pass in the directory containing the data, instead
of a python file to run.