pbio/control: Implement smart coast.
Added `Stop.COAST_SMART` as `then` option for motors. This still coasts the
motor, but it keeps track of the previously used position target. When a new
relative angle command is given (e.g. rotate 90 degrees), it uses that
position as the starting point. This avoids accumulation of errors when using
relative angles in succession