pbio/control: Reset completion state between program runs.
The smart coast option introduces a dependency between sequential
maneuvers, since the next starting point depends on the previous
endpoint. We don't want this dependency to persist when starting a new
program or reinitializing the device.

This is accomplished by resetting the relevant state when initializing
the controller when setting up a servo or drivebase.