pybricks.common.Motor: Add stalled().
A motor can be stalled when control is not active, so it belongs to the
motor class.

By not having to go through the control object, the method is also a bit
easier to use, and it becomes possible to detect stall on Move Hub,
which does not have the Control class.