bricks/stm32: drop firmware.bin targets
Since we use pybricksdev for `make deploy`, we no longer need make
targets to build flash-ready firmware.bin files. Instead, we just
build the firmware.zip file with the firmware-base.bin file and let
pybricksdev take care of the rest.

This is done in the spirit of "there should only be one way to do a
thing".