drv/counter_lpf2: Rework uartdev counter.
The purpose of this driver is to clean up the counter mixed in with
uartdev, which was getting hard to follow.

We will be using a single motor mode going forward. By dropping the
hardcoded multi-mode setup and readout, we can treat motors like any
other LPF2 sensor. This also removes user restrictions on setting
the mode.

This also reduces code size by about 250 bytes.