pbio/trajectory: Use getter for reading endpoint.
When we switch data types, it is safer to use a getter so we can adjust, offset, and scale values as needed, as opposed to reading them from the trajectory structure directly.

It also makes the control code easier to read if you don't know the exact symbols used in the trajectory math.