pbio/trajectory: Use relative motion.
All trajectory math is now evaluated with respect to its starting point.

When using the getter to evaluate the reference, it adds the starting
point to the result, to obtain the same absolute reference as before.

When limiting trajectories to suitable lengths, we can now start
revising the code to use different units while ensuring there are no
overflows after large amount of rotations or time passed.