drv/counter: Return rotations and millidegrees.
The majority of our counter drivers support (or can support) higher
resolution than just one degree. This is a first step in exposing that
functionality.

Because the scale is driver specific, it makes sense to scale in
the drivers instead of returning counts and somehow dealing with
scaling at the higher level. Because of this, we may want to rename this
from a counter to rotation_sensor in the long run.