pbio/control: Revise for increased position resolution.
The mathematical model for the state observer can produce arbitrary
resolution, but its output was truncated to whole degrees. The same
holds for some motors that can output sub-degree resolution.

This might lead to a minor control performance improvement, but it
mainly makes the overal library more consistent and explicit in terms
of units.

Now, applications (servo, drivebase, ...) can define the ratio of the
application units (such as degrees or mm driven) and the internally used
control units (such as millidegrees). This is easier to follow and
document than using platform dependent "counts" as we did until now.

We also drop the use of libfixmath and the expensive conversion
functions (in terms of build size) we used around it.