pbio/drivebase: Change base units and default settings.
Previously, the drive base "distance" was expressed as the sum of
the left and right angles. Now that we use higher resolution
everywhere, we can instead use the average angle without
a large roundoff as a side effect.

This makes the control settings much easier to configure, because
everything has the same control units as single motors.

While we are doing this, we review the settings to pick decent defaults
that work well enough on medium to large sized drivebases.