drv/clock: Add 100us counter.
Compared to the millisecond counter, this gives us just enough
resolution for the controller. This is mainly useful when end users
change the command in a tight loop without waits.

At the same time, it runs much longer than the microsecond counter
before it overflows, which helps with some of the trajectory math
operations.