tools: Add playground notebook.
A Jupyter notebook can often provide an easy way to try out new ideas
without the hassle of setting up a new Python debug configuration or
package.

This commit sets up one such notebook that we can use from time to time,
with full access to the existing venv.

It also comes with an example plot that we can copy and modify to try
out new experiments.