pbio/control: Use scaling function for control gains.
Control gains are in uNm per degree since otherwise their values would
be too small. Position signals use millidegrees, so we have to scale by
a factor of 1000.

This introduces scaling functions both ways, to further reduce the
number of hardcoded scaling operations.

This also fixes a bug where the factor 1000 was missed in the recent
unit conversion, causing integral control to start late or not at all.