pbio/drivebase: Clean up config for SPIKE variant.
There used to be a PBIO_CONFIG_CONTROL_MINIMAL setting to
disable multiple features on Move Hub to save space. After rewriting
much of those features, there's only one left that needs to be disabled.

So we can name it accordingly and use the standard pattern to enable
and disable it. This is only used on SPIKE Prime and SPIKE Essential.