pbio/trajectory: Fix rebasing of infinite trajectories.
When an "infinite" trajectory "expires", we create a new trajectory
from that point onwards. This would lead to bad output during one
sample because the new trajectory was not yet being used.

Also fix several asserts used for range checking.

This also adds a starting point for a test that walks across all points
on the trajectory. This gets computationally prohibitive if we did it
for all permutations, so we'll need to find a way to reduce it.