pbio/trajectory: Fix possible zero division.
For short maneuvers, the speed may get reduced so it doesn't
overshoot the target. In this case, there is no constant-speed
phase since the speed just ramps up and down again.

But we were still computing the duration of the constant speed
phase in all cases, which means a zero division if the speed was
reduced to zero (0/0).

Instead we do this division only if there is a nonzero angle to
travel.

Also fix allowing trj->th1 to be negative. This can be the case if the
initial speed is negative.