drv/ioport_lpf2: set vcc pin to input on shutdown
There seem to still be issues with the some city hubs turning back on
when the power button is released.

This changes the gpio state for the I/O port VCC pin from out low to
input (floating) on shutdown. This seems to fix a similar issue
observed on the move hub as well, so we can drop the quirk for that
platform.

Fixes: https://github.com/pybricks/support/issues/692