drv/imu_lsm6ds3tr_c_stm32: change sensitivity ranges
This changes the accelerometer sensitivity from 2G to 8G and the gyro
sensitivity from 250dps to 1000dps.

The old values were the smallest possible values and the new values are
one step above the mid range values.

For comparison, MINDSTORMS gyro sensors in the past typically have a
sensitivity of 440dps.