pybricks: Use standard way to register modules.
This uses the MP_REGISTER_MODULE macro to register the modules. To make
them part of the pybricks "package", the module names are prefixed with
pybricks_dot_, and the real names with dots are added to the qstrdefs.

This method uses standard MicroPython tools, but it is a bit more
fragile since it relies implicitly on the order of the py/builtinimport.
It works on embedded ports if external imports are disabled. This is
no loss of generality since there aren't any modules or files.

On ev3dev, this almost works as well. Because the builtin modules
take precedence, we have to take care to give them a different name
if there is a Python file with the same name.

Another advantage is that the __init__ method now gets called properly
if it exists, which lets us remove an old workaround on ev3dev.