pbsys/main: Move program download.
This takes the first step towards receiving the program in pbsys.

This makes sense in the long run, and it is useful now for program
storage management.

To make it easy to see what is going on, we make the change in several
steps. In this commit, the program reception process is moved as-is.
This keeps the blocking calls, but drops the MicroPython HAL references.