platform: reset I2C devices at startup
Since hubs can be reset at any time (i.e. by NVIC_SystemReset()),
sometimes this happens in the middle of an I2C transaction, leaving
the remote I2C device in a bad state. Since the MCU resets but not
the remote devices, we need to do a bus reset to get the remote
devices back in a working state. This is done by togging the SCL line
a few times.