bricks/virtualhub: Use attr delegation in package.
This fixes import errors introduced in https://github.com/pybricks/pybricks-micropython/commit/cd16b4c2df0810164254fd52133c6c30e21993da.

With external imports enabled, importing a submodule will try to store
an attribute in the outer module. This can't be done since it is a fixed
table. MicroPython has a feature for delegating to a custom attribute
handler if this happens, which we can use to fix this.

The ev3dev version did not reveal this on CI tests because it still has
a "real" package made with Python files. We should still enable this
feature to handle the situation correctly if/when the file-based package
is dropped.