pbsys: Move program download out of MicroPython.
Just like starting and stopping the program, (down)loading a program is
a task for pbsys. Using a protothread, the download procedure is now
much easier to follow.

Program data is preserved, so that it can be restarted with the button.
This paves the way for supporting persistent programs on all hubs.

By cleaning up the way timeouts work, the user is now less likely to see
a message about an incompatible mpy file when accidentally typing
something in the I/O pane of Pybricks Code.

This also drops the shutdown longjmp. The download thread is cancellable,
and the user program was already cancellable, so we don't need this anymore.

This simplifies the main code and it saves about 500 bytes on Move Hub,
which turns this commit into a slight decrease in code size.