pbsys: Refactor user program start and stop.
Before https://github.com/pybricks/pybricks-micropython/pull/109, the
main application (MicroPython) ran in an infinite loop (goto), so
special hooks were needed to reset and init pbsys prior to running a
new program.

Now, MicroPython runs once before yielding back to pbsys, so these
hooks are no longer needed. Instead, the application can provide a
program stop and stdin event handler alongside the program runner.

This makes the code easier to follow and it makes it possible to
fix https://github.com/pybricks/support/issues/526 by avoiding a full
pbsys reset when dropping from a user program into the REPL.

This also fixes an issue introduced in https://github.com/pybricks/pybricks-micropython/pull/109
where the program would not be stopped on shutdown if the stop button
was disabled.

The only code remaining in pbsys/user_program is about stopping
programs, so this is renamed to program_stop, to be consistent with
program_load.