pbsys/main: Allow graceful exit of application.
This ensures that system resources do not stop until the program has exited and
user data has been saved.

The shutdown animations keep going until all data is saved, so the user can see
when it is safe to remove the batteries.

This also fixes https://github.com/pybricks/pybricks-micropython/commit/5a837734a5970cd6b5c5d4de48f1b5c4e764ef06
in that the light pattern was only visible when the power button was held or
the hub was being charged.

This also fixes the hub not shutting off if the stop button was disabled.