bricks/stm32: Set user flash to useful length.
Before https://github.com/pybricks/pybricks-micropython/pull/109 and related
commits, the user script size was limited to about 25% of RAM. The limit on
storage was not explicitly defined.

With the aforementioned updates, the amount of space is explicitly defined and
constrained by the size of the user flash. This commit updates the size of
those regions to make them the same or larger than the previous program size
limit limit, to ensure there are no regressions in the user experience.

The size could be made a bit larger, but we'll keep the values on the lower end
so we'll still have room for the firmware to grow. See
https://github.com/pybricks/support/issues/713 for an overview and the full
consideration.

For powered up hubs, this means that the user data partially overlaps with the
region scanned by the bootloader to compute the checksum. This commit adds a
checksum complement to the user data map to compensate.