bricks/stm32: Fix PB_FIRMWARE_MAX_SIZE not used.
These were not used anywhere. Instead, the length was inferred from the flash
region size.

We still need a value quite like this, however, since the size scanned by the
bootloader may differ from the flash region defined for the firmware.

This adds a "checksum-scan-size" property for that purpose.

Note: For the moment, it is aliased to "max-firmware-size". We need to
fix this when we update the metadata everywhere at once.