pbsys/program_load: Fix counting of previous checksum complement.
The existing value should not be counted. It is set to zero to make this
explicit, but this got optimized out. Subtracting it after the loop as
part of the final calculation was also optimized out, so make it
volatile instead.

This fixes the checksum being wrong the second time a program was saved.