tools/metadata: update version to v2.0.0
This changes the firmware.metadata.json format to v2.0.0. This is a
breaking change since it drops support for appending a main.py file
to the firmware blob.

Also simplify some names while we are making breaking changes.

Support for this for this format was added in pybricksdev v1.0.0-alpha.31
so the pybricksdev dependency is updated.