sys/program_load: move RAM allocation to C code
This makes the allocation of the user RAM entirely local to the
pbsys_program_load module instead of having some config mixed
into the linker scripts.

This also lets us tune the code/heap ratio on a per-hub basis instead
of hard-coding only 3K for heap.

Instead of a .user_ram section, we now have a more generic .noinit
section in RAM that is not zeroed during early boot. This is used for
the user RAM since it is restored from nonvolatile memory on startup.