drv/block_device_w25qxx_stm32: move size out of linker script
This driver is for external flash, so it doesn't make sense to put
some of it's config in the linker script (which only deals with
internal flash). This moves the config from the linker script to the
platform.c file where the rest of the config already is.

Also changed a couple while loops to for loops to make the code more
idiomatic while we are touching this.