pbio/control: Use speed as numerical derivative.
For PID control, the estimated speed provides the best result, but the
speed value is not representative under stall conditions.

To get a useful value for stall detection and completion detection, we
can numerically differentiate the measured angle. Such a value used to
be provided by the motor, but this value is not provided in the higher
resolution angle mode that we recently started using.