bricks/common: Split common and stm32 files.
So far, only the stm32 builds share makefiles and configuration files, and a main application. All other builds have their own, which means that we need to edit 3*2 or more files just to add one module to pbio or the pybricks package. The main application (stm32/nxt/ev3rt) easily gets out of sync as well.

This is a first step towards organizing config and make files so that they can be common to all builds.

This commit changes only file names and include paths; no code is changed.