bricks/common: Split out config options.
This will make them useful for non-stm32 builds like nxt or ev3rt.

At the moment, the common values are equal to what was used for the stm32 builds. We could use a second option level to enable more options in the unix builds if those eventually share this file too.

Also fix indentation while we are working in these files.