Makefile: Exclude ev3dev-armel from "all" target.
The all target is not used for CI and mainly serves as a convenient way
to build everything locally quickly as a check prior to comitting.

The ev3dev-armel target requires a bit more setup while ev3dev-host
is sufficient for local sanity checks anyway, so let's drop it here.