pbio/servo: Restore user speed as numerical derivative.
Restored the `Motor.speed()` method and `DriveBase` equivalent to
provide speed as a numerical derivative of the motor position, instead
of a model-based estimate. For most end-user cases, this is a more
intuitive result because this value is not affected by mechanical load.

This is fairly similar to the value previously reported by the motor
when running in multi-mode functionality.

The model-based value is only used internally for PID control,
where it is used to provide damping.

Since the model-based value was never used for user speed values
in any releases (only in v3.2bN), this is not a breaking change.