pbio/main: replace pbio_do_one_event with pbio_process_events
We always want to process all pending events, so it makes sense to have
a function that does this instead of only processing one event and
having to use a while loop every time we call it. In some cases, we
did't have the while loop, so in those cases, this should make it more
likely that events are handled in a timely fashion.

In the pb_poll function for stm32, we also add an early exit for cases
where events were handled similar to the virtualhub implementation
which should eliminate some unnecessary waiting.