pbsys/program_load: Reserve space for user data.
A popular feature request is to be able to store some runtime data on the hub, such as the last-used speed of a train.

This commit reserves some space for this purpose so we can avoid breaking user programs stored on (external) flash when we add a user API later.

See https://github.com/pybricks/support/issues/85