pybricks.robotics.DriveBase: Add done() method.
This inverts the busy() method that was never officially released. Busy is not the correct term in the context of hold.

Instead of:
```
while not drive_base.distance_control.done() or not drive_base.heading_control.done():
    wait(10)
```
You can now simply do:
```
while not drive_base.done()
    wait(10)
```

This is useful when doing drivebase commands and other tasks in parallel.