pybricks.common.Motor: Add done() method.
The DriveBase.done(), DriveBase.stalled(), and Motor.stalled() all have
technical reasons to have them as methods instead of getting this info
through the control object.

To be consistent, it makes sense to have Motor.done() as well. This also
has the benefit of being enabled on Boost which does not have the
control attribute.