pbio/test/trajectory: Fix skipping start of trajectory.
When walking the trajectory we have to skip the first point because there is no previous point to compare to. We already skipped 0, but there can be new matching starting points when a long trajectory rebases itself.

This was discovered when adding a missing assert for positive time, which is added in this commit as well.