pbsys/program_stop: Remove shutdown one-shot.
Since the program stop just schedules something, this does not need to be protected by a one-shot.

In fact, this is slightly safer because the mp_pending_exception could theoretically be overwritten by something other than SystemAbort before it gets handled by the VM. Now it will just keep scheduling the SystemAbort.