pybricks: Use normal mp_obj_type_t types.
We need to have unmodified mp_obj_type_t types because type
definitions are going to be very different in MicroPython 1.20. The
modified type also made code harder to read whenever the type was
need in other code (accessed as _type.type).

This reverts 5526a7183a5ee9c2e60c5c10823ce218d4291ab5. Unlike what is suggested
there, adding a special entry in the locals_dict is not a hack but also
used in other parts of MicroPython like MP_MODULE_ATTR_DELEGATION_ENTRY.

This is only a slight increase in build size (+40) because this variant
does not treat the hub type separately. We can still do that if we want
to save some bytes by making it slightly less generic.