pybricks.common.System: Change type definition to skip self check.
All of the methods in the System class are implemented as staticmethods
so self was not being passed, which causes the self-type check to fail.

This works around it without changing the methods by implementing the
System instance as a module, just like we do for the battery.

Fixes: https://github.com/pybricks/support/issues/837

See also https://github.com/pybricks/support/issues/840 for the cleanup
to make the implementation consistent across modules.